refactor: remove layering violations between features/ and Discord deps
features/ no longer imports interactions anywhere, and extensions/ no
longer imports pymongo:

- features/coloc/api_client.py: download_image() now returns raw bytes;
  the zunivers events mixin wraps them in interactions.File itself.
- features/coloc/utils.py: keeps only the pure helpers (date parsing,
  item formatting, URL checks); the five embed builders move to the new
  extensions/zunivers/embeds.py, matching the embeds.py convention used
  by mdi/vlrgg/zevent.
- features/reactionroles/builders.py moves to
  src/discord_ext/rolemenus.py (the documented home for shared
  interactions-dependent UI helpers) so the slash-command extension and
  the WebUI route keep rendering menus identically; the model import is
  TYPE_CHECKING-only so src/discord_ext gains no runtime dependency on
  features/.
- features/xp/repository.py now translates pymongo errors into
  src.core.errors.DatabaseError (and absorbs the benign DuplicateKeyError
  insert race), so the xp extension mixins catch DatabaseError instead of
  importing pymongo.

CLAUDE.md: drop the known-violations caveat and document the
DatabaseError translation + shared-renderer conventions instead.

ruff, mypy, and the 73-test suite pass.
